removing old versions of cross resonance calibration and reordering folder. Also creating Parent sharedf classed for Data, Results and Parametrers used for each Hamiltonian Tomography experiment.
