+++ src/qibocal/protocols/two_qubit_interaction/cross_resonance/__init__.py
@@ -1,22 +1,15 @@
-from .amplitude import cross_resonance_amplitude
-from .hamiltonian_tomography import (
-    hamiltonian_tomography_canc_amplitude,
-    hamiltonian_tomography_canc_phase,
-    hamiltonian_tomography_cr_amplitude,
-    hamiltonian_tomography_cr_length,
-)
-from .length import cross_resonance_length
+from .cancellation_amplitude import hamiltonian_tomography_canc_amplitude
+from .cancellation_phase import hamiltonian_tomography_canc_phase
+from .control_amplitude import hamiltonian_tomography_cr_amplitude
+from .length import hamiltonian_tomography_cr_length
 from .off_res_crosstalk_amp import cr_crosstalk_amplitude
 from .off_res_crosstalk_length import cr_crosstalk_length
 
 __all__ = [
-    "cross_resonance_amplitude",
-    "hamiltonian_tomography_cr_length",
     "hamiltonian_tomography_cr_amplitude",
-    "cross_resonance_length",
-    "cross_resonance_cr_amplitude",
-    "hamiltonian_tomography_canc_amplitude",
+    "hamiltonian_tomography_cr_length",
     "hamiltonian_tomography_canc_phase",
+    "hamiltonian_tomography_canc_amplitude",
     "cr_crosstalk_amplitude",
     "cr_crosstalk_length",
 ]
