+++ src/qibocal/protocols/two_qubit_interaction/cross_resonance/fitting.py
