--- src/qibocal/protocols/two_qubit_interaction/cross_resonance/hamiltonian_tomography/__init__.py
@@ -1,11 +1,0 @@
-from .cancellation_amplitude import hamiltonian_tomography_canc_amplitude
-from .cancellation_phase import hamiltonian_tomography_canc_phase
-from .control_amplitude import hamiltonian_tomography_cr_amplitude
-from .length import hamiltonian_tomography_cr_length
-
-__all__ = [
-    "hamiltonian_tomography_cr_amplitude",
-    "hamiltonian_tomography_cr_length",
-    "hamiltonian_tomography_canc_phase",
-    "hamiltonian_tomography_canc_amplitude",
-]
